+++ R/print.R
@@ -5,8 +5,9 @@ as_fragment <- function(x, page = page_fluid) {
   x
 }
 
-as_page <- function(x) {
+as_page <- function(x, theme = bs_theme()) {
   class(x) <- c("bslib_page", class(x))
+  attr(x, "bs_theme") <- theme
   x
 }
 
@@ -25,8 +26,13 @@ print.bslib_fragment <- function(x, ...) {
 #' @rdname html-browse
 #' @export
 print.bslib_page <- function(x, ...) {
+  old_global <- bs_global_get()
+  bs_global_set(attr(x, "bs_theme", exact = TRUE))
+  on.exit(bs_global_set(old_global))
+
   if (interactive()) {
     x <- htmltools::browsable(x)
   }
+
   invisible(NextMethod())
 }