+++ inst/examples/value_box/deploy.R
@@ -1,6 +1,6 @@
 rsconnect::deployApp(
   rprojroot::find_package_root_file("inst/examples/value_box"),
   appName = "value_box",
-  account = "testing-apps",
+  account = "bslib",
   forceUpdate = TRUE
 )